internals cleanup
