+++ src/arch/x86_64.rs
@@ -1,4 +1,4 @@
-//! Intrinsics for the [x86_64](https://en.wikipedia.org/wiki/X86-64) processor family.
+//! Intrinsics for the [`x86_64`](https://en.wikipedia.org/wiki/X86-64) processor family.
 
 use super::*;
