+++ src/lib.rs
@@ -12,6 +12,9 @@ use core::{
 
 pub mod arch;
 
+mod marker;
+pub use marker::*;
+
 /// Fiddly to use, but totally gets you the minimum without branching.
 ///
 /// Works for any integral type.
@@ -30,20 +33,6 @@ macro_rules! branchless_max {
   ($x:ident, $y:ident, $u:ty) => {
     $x ^ (($x ^ $y) & (<$u>::wrapping_neg(($x < $y) as $u)))
   };
-}
-
-/// Implements an unsafe marker trait on an array type if the element type
-/// also supports that marker trait.
-///
-/// Syntax:
-///
-/// ```txt
-/// impl_unsafe_marker_for_array!(TraitName, 0, 1, 2, 3, ...);
-/// ```
-macro_rules! impl_unsafe_marker_for_array {
-  ( $marker:ident , $( $n:expr ),* ) => {
-    $(unsafe impl<T> $marker for [T; $n] where T: $marker {})*
-  }
 }
 
 /// Wrapper inner value to a minimum of 2.
@@ -86,115 +75,21 @@ pub struct Align16<T>(pub T);
 #[repr(align(32))]
 pub struct Align32<T>(pub T);
 
-/// A trait for types that can be safely made with [zeroed](core::mem::zeroed).
+/// Re-interprets `&T` as `&[u8]`.
 ///
-/// ## Safety
-///
-/// * Your type must be inhabited (eg: no
-///   [Infallible](core::convert::Infallible)).
-/// * Your type must be allowed to be an "all zeroes" bit pattern (eg: no
-///   [NonNull<T>](core::ptr::NonNull)).
-pub unsafe trait Zeroable: Sized {
-  /// Calls [zeroed](core::mem::zeroed).
-  ///
-  /// This is a trait method so that you can write `MyType::zeroed()` in your
-  /// code. It is a contract of this trait that if you implement it on your type
-  /// you _must not_ override this method. In the future this trait will become
-  /// a `#[marker]` trait so that this is absolute. Until then just don't do it.
-  fn zeroed() -> Self {
-    unsafe { core::mem::zeroed() }
-  }
+/// Any ZST becomes an empty slice, and in that case the pointer value of that
+/// empty slice might not match the pointer value of the input reference.
+pub fn bytes_of<T: Pod>(t: &T) -> &[u8] {
+  try_cast_slice::<T, u8>(core::slice::from_ref(t)).unwrap_or(&[])
 }
-unsafe impl Zeroable for () {}
-unsafe impl Zeroable for bool {}
-unsafe impl Zeroable for char {}
-unsafe impl Zeroable for u8 {}
-unsafe impl Zeroable for i8 {}
-unsafe impl Zeroable for u16 {}
-unsafe impl Zeroable for i16 {}
-unsafe impl Zeroable for u32 {}
-unsafe impl Zeroable for i32 {}
-unsafe impl Zeroable for u64 {}
-unsafe impl Zeroable for i64 {}
-unsafe impl Zeroable for usize {}
-unsafe impl Zeroable for isize {}
-unsafe impl Zeroable for u128 {}
-unsafe impl Zeroable for i128 {}
-unsafe impl Zeroable for f32 {}
-unsafe impl Zeroable for f64 {}
-unsafe impl Zeroable for Option<NonZeroI8> {}
-unsafe impl Zeroable for Option<NonZeroI16> {}
-unsafe impl Zeroable for Option<NonZeroI32> {}
-unsafe impl Zeroable for Option<NonZeroI64> {}
-unsafe impl Zeroable for Option<NonZeroI128> {}
-unsafe impl Zeroable for Option<NonZeroIsize> {}
-unsafe impl Zeroable for Option<NonZeroU8> {}
-unsafe impl Zeroable for Option<NonZeroU16> {}
-unsafe impl Zeroable for Option<NonZeroU32> {}
-unsafe impl Zeroable for Option<NonZeroU64> {}
-unsafe impl Zeroable for Option<NonZeroU128> {}
-unsafe impl Zeroable for Option<NonZeroUsize> {}
-unsafe impl<T> Zeroable for *mut T {}
-unsafe impl<T> Zeroable for *const T {}
-unsafe impl<T> Zeroable for Option<NonNull<T>> {}
-unsafe impl<T> Zeroable for PhantomData<T> where T: Zeroable {}
-impl_unsafe_marker_for_array!(
-  Zeroable, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,
-  24, 25, 26, 27, 28, 29, 30, 31, 32, 48, 64, 96, 128, 256, 512, 1024
-);
 
-/// A marker trait for "plain old data".
+/// Re-interprets `&mut T` as `&mut [u8]`.
 ///
-/// The point of this trait is that once something is marked "plain old data"
-/// you can really go to town with the bit fiddling and bit casting. Therefore,
-/// it's a relatively strong claim to make about a type. Do not add this to your
-/// type casually.
-///
-/// ## Safety
-///
-/// * The type must be inhabited (eg: no
-///   [Infallible](core::convert::Infallible)).
-/// * The type must allow any bit pattern (eg: no `bool` or `char`).
-/// * The type must not contain any padding bytes (eg: no `(u8, u16)`).
-/// * A struct needs to be `repr(C)`, or a `repr(transparent)` wrapper around a
-///   `Pod` type.
-pub unsafe trait Pod: Zeroable + Copy {}
-
-unsafe impl Pod for () {}
-unsafe impl Pod for u8 {}
-unsafe impl Pod for i8 {}
-unsafe impl Pod for u16 {}
-unsafe impl Pod for i16 {}
-unsafe impl Pod for u32 {}
-unsafe impl Pod for i32 {}
-unsafe impl Pod for u64 {}
-unsafe impl Pod for i64 {}
-unsafe impl Pod for usize {}
-unsafe impl Pod for isize {}
-unsafe impl Pod for u128 {}
-unsafe impl Pod for i128 {}
-unsafe impl Pod for f32 {}
-unsafe impl Pod for f64 {}
-unsafe impl Pod for Option<NonZeroI8> {}
-unsafe impl Pod for Option<NonZeroI16> {}
-unsafe impl Pod for Option<NonZeroI32> {}
-unsafe impl Pod for Option<NonZeroI64> {}
-unsafe impl Pod for Option<NonZeroI128> {}
-unsafe impl Pod for Option<NonZeroIsize> {}
-unsafe impl Pod for Option<NonZeroU8> {}
-unsafe impl Pod for Option<NonZeroU16> {}
-unsafe impl Pod for Option<NonZeroU32> {}
-unsafe impl Pod for Option<NonZeroU64> {}
-unsafe impl Pod for Option<NonZeroU128> {}
-unsafe impl Pod for Option<NonZeroUsize> {}
-unsafe impl<T> Pod for *mut T {}
-unsafe impl<T> Pod for *const T {}
-unsafe impl<T> Pod for Option<NonNull<T>> {}
-unsafe impl<T> Pod for PhantomData<T> where T: Pod {}
-impl_unsafe_marker_for_array!(
-  Pod, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24,
-  25, 26, 27, 28, 29, 30, 31, 32, 48, 64, 96, 128, 256, 512, 1024
-);
+/// Any ZST becomes an empty slice, and in that case the pointer value of that
+/// empty slice might not match the pointer value of the input reference.
+pub fn bytes_of_mut<T: Pod>(t: &mut T) -> &mut [u8] {
+  try_cast_slice_mut::<T, u8>(core::slice::from_mut(t)).unwrap_or(&mut [])
+}
 
 /// The things that can go wrong when casting a slice.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -211,6 +106,16 @@ pub enum SliceCastError {
   /// accordingly. If the output slice wouldn't be a whole number of elements
   /// then the conversion fails.
   OutputSliceWouldHaveSlop,
+}
+
+/// As [try_cast_slice](try_cast_slice), but unwraps the result for you.
+pub fn cast_slice<A: Pod, B: Pod>(a: &[A]) -> &[B] {
+  try_cast_slice(a).unwrap()
+}
+
+/// As [try_cast_slice_mut](try_cast_slice_mut), but unwraps the result for you.
+pub fn cast_slice_mut<A: Pod, B: Pod>(a: &[A]) -> &[B] {
+  try_cast_slice(a).unwrap()
 }
 
 /// Try to convert a slice of one type into another.
@@ -238,20 +143,15 @@ pub fn try_cast_slice<A: Pod, B: Pod>(a: &[A]) -> Result<&[B], SliceCastError> {
     Ok(unsafe { core::slice::from_raw_parts(a.as_ptr() as *const B, a.len()) })
   } else if size_of::<A>() == 0 || size_of::<B>() == 0 {
     Err(SliceCastError::CantConvertBetweenZSTAndNonZST)
-  } else if core::mem::size_of_val(a) % size_of::<B>() != 0 {
-    Err(SliceCastError::OutputSliceWouldHaveSlop)
-  } else {
+  } else if core::mem::size_of_val(a) % size_of::<B>() == 0 {
     let new_len = core::mem::size_of_val(a) / size_of::<B>();
     Ok(unsafe { core::slice::from_raw_parts(a.as_ptr() as *const B, new_len) })
+  } else {
+    Err(SliceCastError::OutputSliceWouldHaveSlop)
   }
 }
 
-/// As [try_cast_slice](try_cast_slice), but unwraps the result for you.
-pub fn cast_slice<A: Pod, B: Pod>(a: &[A]) -> &[B] {
-  try_cast_slice(a).unwrap()
-}
-
-/// As [try_cast_slice](try_cast_slice), but mut.
+/// As [try_cast_slice](try_cast_slice), but `mut`.
 pub fn try_cast_slice_mut<A: Pod, B: Pod>(a: &mut [A]) -> Result<&mut [B], SliceCastError> {
   // Note(Lokathor): everything with `align_of` and `size_of` will optimize away
   // after monomorphization.
@@ -261,31 +161,10 @@ pub fn try_cast_slice_mut<A: Pod, B: Pod>(a: &mut [A]) -> Result<&mut [B], Slice
     Ok(unsafe { core::slice::from_raw_parts_mut(a.as_ptr() as *mut B, a.len()) })
   } else if size_of::<A>() == 0 || size_of::<B>() == 0 {
     Err(SliceCastError::CantConvertBetweenZSTAndNonZST)
-  } else if core::mem::size_of_val(a) % size_of::<B>() != 0 {
-    Err(SliceCastError::OutputSliceWouldHaveSlop)
-  } else {
+  } else if core::mem::size_of_val(a) % size_of::<B>() == 0 {
     let new_len = core::mem::size_of_val(a) / size_of::<B>();
     Ok(unsafe { core::slice::from_raw_parts_mut(a.as_ptr() as *mut B, new_len) })
+  } else {
+    Err(SliceCastError::OutputSliceWouldHaveSlop)
   }
-}
-
-/// As [try_cast_slice_mut](try_cast_slice_mut), but unwraps the result for you.
-pub fn cast_slice_mut<A: Pod, B: Pod>(a: &[A]) -> &[B] {
-  try_cast_slice(a).unwrap()
-}
-
-/// Re-interprets a reference as a byte slice reference.
-///
-/// Any ZST becomes an empty slice, and in that case the pointer value of that
-/// empty slice might not match the pointer value of the input reference.
-pub fn bytes_of<T: Pod>(t: &T) -> &[u8] {
-  try_cast_slice::<T, u8>(core::slice::from_ref(t)).unwrap_or(&[])
-}
-
-/// Re-interprets a mut reference as a mut byte slice reference.
-///
-/// Any ZST becomes an empty slice, and in that case the pointer value of that
-/// empty slice might not match the pointer value of the input reference.
-pub fn bytes_of_mut<T: Pod>(t: &mut T) -> &mut [u8] {
-  try_cast_slice_mut::<T, u8>(core::slice::from_mut(t)).unwrap_or(&mut [])
 }
